requirements.txt
pymupdf
sentence-transformers
faiss-cpu
numpy
tqdm
